SERVICES-1423 user energy cleanup
- remove unnecessary user service
- code cleanup on user update energy transaction
